test(fixtures): refactor fixtures
