+++ src/tests/endpoints/test_code.py
@@ -1,38 +1,8 @@
 from typing import Any, Dict, Union
 
 import pytest
-import pytest_asyncio
 from httpx import AsyncClient
 from sqlmodel.ext.asyncio.session import AsyncSession
-
-from app.models import User
-
-USER_TABLE = [
-    {
-        "id": 1,
-        "provider_user_id": 123,
-        "login": "first_login",
-        "hashed_password": "hashed_first_pwd",
-        "scope": "admin",
-        "created_at": "2024-02-23T08:18:45.447773",
-    },
-    {
-        "id": 2,
-        "provider_user_id": 456,
-        "login": "second_login",
-        "hashed_password": "hashed_second_pwd",
-        "scope": "user",
-        "created_at": "2024-02-23T08:18:45.447774",
-    },
-]
-
-
-@pytest_asyncio.fixture(scope="function")
-async def code_session(async_session: AsyncSession, monkeypatch):
-    for entry in USER_TABLE:
-        async_session.add(User(**entry))
-    await async_session.commit()
-    yield async_session
 
 
 @pytest.mark.parametrize(
@@ -83,7 +53,7 @@ async def code_session(async_session: AsyncSession, monkeypatch):
 @pytest.mark.asyncio()
 async def test_chat(
     async_client: AsyncClient,
-    code_session: AsyncSession,
+    user_session: AsyncSession,
     user_idx: Union[int, None],
     payload: Dict[str, Any],
     status_code: int,
@@ -91,7 +61,7 @@ async def test_chat(
 ):
     auth = None
     if isinstance(user_idx, int):
-        auth = await pytest.get_token(USER_TABLE[user_idx]["id"], USER_TABLE[user_idx]["scope"].split())
+        auth = await pytest.get_token(pytest.user_table[user_idx]["id"], pytest.user_table[user_idx]["scope"].split())
 
     response = await async_client.post("/code/chat", json=payload, headers=auth)
     assert response.status_code == status_code, print(response.__dict__)
