+++ src/tests/endpoints/test_login.py
@@ -2,41 +2,8 @@ from typing import Any, Dict, Union
 from urllib.parse import parse_qs, urlparse
 
 import pytest
-import pytest_asyncio
 from httpx import AsyncClient
 from sqlmodel.ext.asyncio.session import AsyncSession
-
-from app.api.api_v1.endpoints import login, users
-from app.models import User
-
-USER_TABLE = [
-    {
-        "id": 1,
-        "provider_user_id": 123,
-        "login": "first_login",
-        "hashed_password": "hashed_first_pwd",
-        "scope": "user",
-        "created_at": "2024-02-23T08:18:45.447773",
-    },
-    {
-        "id": 2,
-        "provider_user_id": 456,
-        "login": "second_login",
-        "hashed_password": "hashed_second_pwd",
-        "scope": "user",
-        "created_at": "2024-02-23T08:18:45.447774",
-    },
-]
-
-
-@pytest_asyncio.fixture(scope="function")
-async def login_session(async_session: AsyncSession, monkeypatch):
-    for entry in USER_TABLE:
-        async_session.add(User(**entry))
-    await async_session.commit()
-    monkeypatch.setattr(login, "verify_password", pytest.mock_verify_password)
-    monkeypatch.setattr(users, "hash_password", pytest.mock_hash_password)
-    yield async_session
 
 
 @pytest.mark.parametrize(
@@ -49,7 +16,7 @@ async def login_session(async_session: AsyncSession, monkeypatch):
 @pytest.mark.asyncio()
 async def test_login_with_github_token(
     async_client: AsyncClient,
-    login_session: AsyncSession,
+    user_session: AsyncSession,
     payload: Dict[str, Any],
     status_code: int,
     status_detail: Union[str, None],
@@ -72,7 +39,7 @@ async def test_login_with_github_token(
 @pytest.mark.asyncio()
 async def test_login_with_creds(
     async_client: AsyncClient,
-    login_session: AsyncSession,
+    user_session: AsyncSession,
     payload: Dict[str, Any],
     status_code: int,
     status_detail: Union[str, None],
@@ -85,7 +52,7 @@ async def test_login_with_creds(
         response_json = response.json()
         assert response_json["token_type"] == "bearer"  # noqa: S105
         assert isinstance(response_json["access_token"], str)
-        assert len(response_json["access_token"]) == 143
+        assert len(response_json["access_token"]) == 144
 
 
 @pytest.mark.parametrize(
@@ -100,7 +67,7 @@ async def test_login_with_creds(
 @pytest.mark.asyncio()
 async def test_request_github_token_from_code(
     async_client: AsyncClient,
-    login_session: AsyncSession,
+    user_session: AsyncSession,
     payload: Dict[str, Any],
     status_code: int,
     status_detail: Union[str, None],
@@ -123,7 +90,7 @@ async def test_request_github_token_from_code(
 @pytest.mark.asyncio()
 async def test_authorize_github(
     async_client: AsyncClient,
-    login_session: AsyncSession,
+    user_session: AsyncSession,
     scope: str,
     redirect_uri: str,
     status_code: int,
